% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt.R
\name{melt_reports}
\alias{melt_reports}
\title{Melt sections of a quaqc report into a data.frame.}
\usage{
melt_reports(report, section = c("bam_stats", "overview_unfilt",
  "overview_filt", "nucl_stats", "nucl_addn", "peak_stats", "tss_stats",
  "tss_pileup", "aln_hist", "frag_hist", "gc_hist", "depth_hist", "tss_pileup",
  "genome"), use.basename = FALSE, normalize.tss = c("no", "bkg", "rpm"),
  normalize.hist = c("no", "proportion", "max"))
}
\arguments{
\item{report}{A \code{quaqc} object.}

\item{section}{The \code{quaqc} object subsection to melt.}

\item{use.basename}{Whether to use the \code{\link[base:basename]{base::basename()}} function
on the sample names.}

\item{normalize.tss}{How to normalize the TSS pileup. "no": Keep
the signal as the average number of reads per window. "bkg":
Calculate the signal relative to the background (the first 25\%
of the window). "rpm": Convert to reads per million.}

\item{normalize.hist}{How to normalize the alignment size,
fragment size, GC percent, and read depth histograms.
"no": Keep as the total number of reads per bin. "proportion":
Divide by the sum of reads across all windows. "max": Divide
by the max bin count.}
}
\value{
A \code{data.frame} with varying columns based on the
section being melted.
}
\description{
The quaqc report class type in R is divided into lists of lists,
which can require additional manipulation. This function will
"melt" these individual sections into \code{data.frame} objects.
}
\examples{
report.file <- system.file("extdata", "report.json.gz", package = "quaqcr")
report <- parse_quaqc_file(report.file)
melt_reports(report, "overview_filt")

}
\seealso{
\code{\link[=quaqc]{quaqc()}}, \code{\link[=parse_quaqc_file]{parse_quaqc_file()}}
}
\author{
Benjamin Jean-Marie Tremblay, \email{benjmtremblay@gmail.com}
}
