% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprint.R
\name{footprint}
\alias{footprint}
\title{Get transcription factor footprints with quaqc.}
\usage{
footprint(target.motifs, bkg.motifs = NULL, bam.files, tss.size = 501,
  tss.qlen = 1, tss.tn5 = TRUE, fast = TRUE, nfr = TRUE, ...)
}
\arguments{
\item{target.motifs}{Either (1) a filename of a BED file containing
target motif positions, or (2) a \code{GRanges} object from the
\pkg{GenomicRanges} package.}

\item{bkg.motifs}{Either (1) a filename of a BED file containing
target motif positions, or (2) a \code{GRanges} object from the
\pkg{GenomicRanges} package. (Optional.)}

\item{bam.files}{Character vector of BAM file names. Must be coordinate
sorted. If no index file can be found, \code{quaqc} will generate them.}

\item{tss.size}{Integer, size of the TSS region for pileup.}

\item{tss.qlen}{Integer, resize reads (centered on the 5-prime end for pileup.}

\item{tss.tn5}{Logical, shift 5-prime end coordinates +4/-5 bases for pileup.}

\item{fast}{Logical, turn on fast mode.}

\item{nfr}{Logical, turn on NFR mode.}

\item{...}{See \code{\link[=quaqc]{quaqc()}}.}
}
\value{
A list of length 2. The first entry contains a \code{data.frame} with
the fooprint pileups, and the second entry contains the quaqc reports.
}
\description{
The TSS pileup feature of quaqc can instead be used to get single
base resolution transcription factor footprints from ATAC-seq data.
This function provides a convenient wrapper around such
functionality. Target regions can be compared to unbound or background
regions. (Note that no Tn5 bias correction is applied.)
}
\seealso{
\code{\link[base:system2]{base::system2()}}, \code{\link[=quaqc]{quaqc()}}
}
\author{
Benjamin Jean-Marie Tremblay, \email{benjmtremblay@gmail.com}
}
