% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{parse_quaqc}
\alias{parse_quaqc}
\alias{parse_quaqc_file}
\title{Parse a JSON quaqc report file.}
\usage{
parse_quaqc(json.text)

parse_quaqc_file(json.file)
}
\arguments{
\item{json.text}{The JSON report as a character vector.}

\item{json.file}{The JSON report filename.}
}
\value{
A \code{quaqc}-class object.
}
\description{
Parse the output of \code{quaqc --json} into a easier to use
\code{quaqc}-class object within R.
}
\details{
Note that the word 'effective' refers to reads which are visible to
quaqc within target regions or outside blacklisted regions, as well
as reads associated with any specified target read groups.
}
\examples{
\dontrun{
f <- gzfile("report.json.gz", "rt")
json <- jsonlite::fromJSON(readLines(f))
close(f)
report <- parse_quaqc(json)
}

}
\author{
Benjamin Jean-Marie Tremblay, \email{benjmtremblay@gmail.com}
}
