% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quaqc.R
\name{quaqc}
\alias{quaqc}
\title{Run quaqc from within R.}
\usage{
quaqc(bam.files, mitochondria = NULL, plastids = NULL, peaks = NULL,
  tss = NULL, target.names = NULL, target.list = NULL,
  blacklist = NULL, rg.names = NULL, rg.list = NULL,
  use.secondary = FALSE, use.nomate = FALSE, use.dups = FALSE,
  use.chimeric = FALSE, use.dovetails = FALSE, no.se = FALSE,
  mapq = NULL, min.qlen = NULL, min.flen = NULL, max.qlen = NULL,
  max.flen = NULL, use.all = FALSE, max.depth = NULL, max.qhist = NULL,
  max.fhist = NULL, tss.size = NULL, tss.qlen = NULL, tss.tn5 = FALSE,
  omit.gc = FALSE, omit.depth = FALSE, fast = FALSE, lenient = FALSE,
  strict = FALSE, nfr = FALSE, nbr = FALSE, footprint = FALSE,
  chip = FALSE, output.dir = NULL, output.ext = NULL, no.output = TRUE,
  json = "-", keep = FALSE, keep.dir = NULL, keep.ext = NULL,
  threads = NULL, title = NULL, continue = FALSE, verbose = 1,
  timeout = 0, env = character(), stderr.file = "",
  bin = getOption("quaqc.bin"))
}
\arguments{
\item{bam.files}{Character vector of BAM file names. Must be coordinate
sorted. If no index file can be found, \code{quaqc} will generate them.}

\item{mitochondria}{Character vector of mitochondria names. Provide \code{""}
to clear the defaults.}

\item{plastids}{Character vector of plastid names. Provide \code{""} to clear
the defaults.}

\item{peaks}{Either (1) a filename of a BED file containing peaks, or (2) a
\code{GRanges} object from the \pkg{GenomicRanges} package.}

\item{tss}{Either (1) a filename of a BED file containing TSSs, or (2) a
\code{GRanges} object from the \pkg{GenomicRanges} package.}

\item{target.names}{Character vector of sequence names to restrict \code{quaqc}.}

\item{target.list}{Either (1) a filename of a BED file containing ranges
to restrict \code{quaqc}, or (2) a \code{GRanges} object from the \pkg{GenomicRanges}
package.}

\item{blacklist}{Either (1) a filename of a BED file containing blacklist
ranges, or (2) a \code{GRanges} object from the \pkg{GenomicRanges}
package.}

\item{rg.names}{A character vector of read group (RG) names to restrict
\code{quaqc}.}

\item{rg.list}{Filename of a text file containing read group (RG) names to
restrict \code{quaqc}, one name per line.}

\item{use.secondary}{Logical, allow secondary alignments.}

\item{use.nomate}{Logical, allow PE reads when the mate does not align
properly.}

\item{use.dups}{Logical, allow duplicate reads.}

\item{use.chimeric}{Logical, allow supplemental or chimeric alignments.}

\item{use.dovetails}{Logical, allow dovetailing PE reads.}

\item{no.se}{Logical, discard SE reads.}

\item{mapq}{Integer, min MAPQ score.}

\item{min.qlen}{Integer, min alignment length.}

\item{min.flen}{Integer, min fragment length.}

\item{max.qlen}{Integer, max alignment length.}

\item{max.flen}{Integer, max fragment length.}

\item{use.all}{Logical, discard all filters and keep all reads.}

\item{max.depth}{Integer, max base depth for read depth histogram.}

\item{max.qhist}{Integer, max alignment length for histogram.}

\item{max.fhist}{Integer, max fragment length for histogram.}

\item{tss.size}{Integer, size of the TSS region for pileup.}

\item{tss.qlen}{Integer, resize reads (centered on the 5-prime end for pileup.}

\item{tss.tn5}{Logical, shift 5-prime end coordinates +4/-5 bases for pileup.}

\item{omit.gc}{Logical, omit calculation of read GC content.}

\item{omit.depth}{Logical, omit calculation of read depths.}

\item{fast}{Logical, turn on fast mode.}

\item{lenient}{Logical, turn on lenient mode.}

\item{strict}{Logical, turn on strict mode.}

\item{nfr}{Logical, turn on NFR mode.}

\item{nbr}{Logical, turn on NBR mode.}

\item{footprint}{Logical, turn on footprinting mode.}

\item{chip}{Logical, turn on ChIP-seq mode.}

\item{output.dir}{Name of directory to save QC report if not that of input.}

\item{output.ext}{Filename extension for output files.}

\item{no.output}{Logical, suppress creation of output QC reports. Note that
option is turned on by default when run from \pkg{quaqcr}.}

\item{json}{Filename of JSON file to save combined QC results to. Set to
\code{NULL} to suppress this. The default is to pipe the JSON output directly to
R and not save to a file.}

\item{keep}{Logical, save passing nuclear reads to a new BAM file.}

\item{keep.dir}{Directory name to save filtered BAMs.}

\item{keep.ext}{Extension of filtered BAMs.}

\item{threads}{Integer, number of worker threads. Max one per sample.}

\item{title}{Assign a title to run.}

\item{continue}{Logical, do not return an error and instead continue running
if samples trigger program errors.}

\item{verbose}{Integer, a value from 0 to 2 for the level of program verbosity.}

\item{timeout}{Integer, number of seconds before stopping quaqc. By default
it is allowed to run indefinitely. See \code{\link[base:system2]{base::system2()}}.}

\item{env}{A character vector of environment variables to set when
running quaqc. See \code{\link[base:system2]{base::system2()}}.}

\item{stderr.file}{Filename to save quaqc messages. By default they are
printed in the console.}

\item{bin}{Path to quaqc binary. Alternatively, set \code{options(quaqc.bin)}.
If the binary is present in the current working directory, provide \code{"./quaqc"}.
The default, set when \pkg{quaqcr} is loaded, is to assume the binary is
present in your PATH.}
}
\value{
If nothing is provided to \code{bam.files}, then the help message is
printed to the console and returned as a character vector, invisibly.
Alternatively: if \code{json = NULL} then \code{NULL}, otherwise the JSON
output as parsed by \pkg{jsonlite}.
}
\description{
Interactive wrapper for running quaqc from within R. For a detailed
description of the program, see the manual: execute \code{man quaqc} from the
command line or open \code{doc/quaqc.1.md} in the program folder. For a brief
description of the command parameters, as well as to see default values,
call \code{quaqc()} without any arguments.
}
\examples{
\donttest{
if (nzchar(Sys.which("quaqc"))) {
  ## To check that you are properly linking to the binary and view help:
  quaqc()
}
}

}
\seealso{
\code{\link[base:system2]{base::system2()}}, \code{\link[=parse_quaqc]{parse_quaqc()}}, \code{\link[=parse_quaqc_file]{parse_quaqc_file()}}
}
\author{
Benjamin Jean-Marie Tremblay, \email{benjmtremblay@gmail.com}
}
